WL-3230 Stop trying to delete the old file when replacing
When receiving a PUT request then we shouldn't need to delete the old file,
let's just send the input stream to the existing file.
Currently the `ContentHostingService.editResource` throws a permission
exception when it should throw an id unused exception.

So we test it first with getResource to see if the file exists.

This can be fixed when KNL-1217 is fixed.